requirements.txt
beautifulsoup4==4.12.2
selenium==4.16.0
webdriver-manager==3.8.6
psycopg2-binary==2.9.9
